refactor(ulw-loop): keep types facade stable
